# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ChangeControlPlane API
  version: 0.1.0
  description: >
    Initial version of the ChangeControlPlane control-plane API for
    organizations, projects, teams, services, environments, changes,
    risk assessments, rollout plans, rollout executions, policies,
    audit events, integrations, service accounts, and machine-token flows.
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
paths:
...
  /api/v1/projects:
    get:
      summary: List projects
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectItemResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/v1/integrations/{id}/webhooks/gitlab:
    post:
      summary: Receive GitLab webhook deliveries for a configured integration
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Webhook delivery accepted and recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationSyncRunItemResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: CCP-HMAC
  responses:
    ValidationError:
      description: Validation or malformed request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Missing, invalid, or expired authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Authenticated, but not permitted for the requested organization or resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    CreateProjectRequest:
      type: object
      required: [organization_id, name, slug]
      properties:
        organization_id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        adoption_mode:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    Project:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        organization_id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        adoption_mode:
          type: string
        status:
          type: string
...
    ProjectListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    ProjectItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Project'
...
    IntegrationSyncRun:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        organization_id:
          type: string
        integration_id:
          type: string
        operation:
          type: string
        trigger:
          type: string
        status:
          type: string
        summary:
          type: string
        details:
          type: array
          items:
            type: string
        resource_count:
          type: integer
        external_event_id:
          type: string
        error_class:
          type: string
        scheduled_for:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
...
    IntegrationSyncRunItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/IntegrationSyncRun'
...